pyproject.toml
[project]
name = "docgen"
version = "0.1.0"
description = "Interactive documentation generator for code projects"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "questionary>=2.0.0",
]

[project.scripts]
docgen = "docgen.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]